apps/host-executor/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "team-agent-platform-host-executor"
version = "0.1.0"
description = "Host execution bridge for Team Agent Platform"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["host_executor_app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]
